libft

+++ libft/srcs/ft_is/ft_isinteger.c
@@ -6,7 +6,7 @@
 /*   By: sregnard <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/20 11:47:24 by sregnard          #+#    #+#             */
-/*   Updated: 2019/02/20 17:05:10 by sregnard         ###   ########.fr       */
+/*   Updated: 2019/03/05 16:08:08 by sregnard         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 int	ft_isinteger(char *s)
 {
-	long long	nb;
+	unsigned int	nb;
 
 	if (!s || !ft_isnumber(s))
 			return (0);
